Code improvements and bug fixes

+++ tortoisebeacon/config.go
@@ -2,8 +2,8 @@ package tortoisebeacon
 
 // Config is the configuration of the Tortoise Beacon.
 type Config struct {
-	K           int `mapstructure:"tortoise-beacon-rounds-number"` // number of rounds
-	WakeupDelta int `mapstructure:"tortoise-beacon-wakeup-delta"`  // the wakeup delta after tick
+	K           uint64 `mapstructure:"tortoise-beacon-rounds-number"` // number of rounds
+	WakeupDelta int    `mapstructure:"tortoise-beacon-wakeup-delta"`  // the wakeup delta after tick
 }
 
 // DefaultConfig returns the default configuration for the tortoise beacon.
